Added Display for Session Variable
Added function to flask views.add_to_session to return the contents of the server addons session variable.
Added line to JS file server_qr.js to update the contents of the html file with the session contents.
Added empty div to the search html to display the session contents.